requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
Pillow>=8.3.1
opencv-python>=4.5.3
PyYAML>=5.4.1
tqdm>=4.62.2
clip @ git+https://github.com/openai/CLIP.git